+++ apps/hosting/urls.py
@@ -35,4 +35,6 @@ urlpatterns = [
         name='hosting-deployment-rollback',
     ),
     path('stats', views.StatsView.as_view(), name='hosting-stats'),
+    path('metrics', views.HostingMetricsView.as_view(), name='hosting-metrics'),
+    path('metrics/all', views.HostingGlobalMetricsView.as_view(), name='hosting-metrics-all'),
 ]